rename: 폴더 경로 및 파일명 통일
