+++ src/components/registerForm/expireDate/index.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
-import Input from "src/components/common/Input";
+import Input from "src/components/@common/Input";
 import styled, { css } from "styled-components";
-import ErrorSpan from "src/components/common/ErrorSpan";
-import FormLabel from "src/components/common/FormLabel";
+import ErrorSpan from "src/components/@common/ErrorSpan";
+import FormLabel from "src/components/@common/FormLabel";
 import { ONLY_MONTH_REGEXP } from "src/utils/regexp";
 
 function isValidMonth(month: string) {
